+++ classes/local/participant/data_access/participant_vault_interface.php
@@ -83,7 +83,7 @@ interface participant_vault_interface {
      * @param int       $studentid The student id.
      * @return grade[]  A student grades.
      */
-    public function get_student_assignment_grades(int $courseid, int $studentid);
+    public function get_student_exercises_grades(int $courseid, int $studentid);
 
     /**
      * Get quiz grades for a specific student.
@@ -173,4 +173,34 @@ interface participant_vault_interface {
      * @return  array   The next exercise id and associated course section
      */
     public function get_student_exercise($courseid, $studentid, $next = true);
+
+    /**
+     * Returns the number of attempts for a specific exercise.
+     *
+     * @param   int     The course id
+     * @param   int     The student user id
+     * @param   int     The exercise id to get the number of attempts for
+     * @return  int     The number of attempts for an exercise
+     */
+    public function get_student_exercise_attempts(int $courseid, int $studentid, int $exerciseid);
+
+    /**
+     * Returns the the skill test assessment, which includes all
+     * skill test sections and thier exercises.
+     *
+     * @param   int     The course id
+     * @param   int     The student user id
+     * @param   string  The skill test main section name for looking up exercises
+     * @return  array   The skill test sections
+     */
+    public function get_student_skilltest_assessment(int $courseid, int $studentid, string $skilltestsecname);
+
+    /**
+     * Returns the the skill test assessment subsections (rubrics).
+     *
+     * @param   int     The student user id
+     * @param   int     The skill test section exercise id (assignment)
+     * @return  array   The skill test subsections
+     */
+    public function get_student_skilltest_subsections(int $studentid, int $assignid);
 }