+++ lang/en/local_booking.php
@@ -306,9 +306,15 @@ $string['theoryexamreport']='Theory examination report';
 
 // reports
 $string['aircrafttypelabel'] = 'Aircraft type/class course conducted in';
-$string['candidatedetails'] = 'Candidate\'s Details';
+$string['appeals'] = 'Appeals';
+$string['appealstext'] = 'Any candidate who has failed a test or examination may within 14 days of being notified of his or her failure request that the
+    Training Department determine whether the test or examination was properly conducted. In order to succeed you will have to satisfy the Training Department
+    that the examination or test was not properly conducted. Mere dissatisfaction with the result is not sufficient reason for appeal.';
+$string['blocktimes'] = 'Block times';
+$string['candidatedetails'] = 'Candidate Details';
 $string['candidateflyinghours'] = 'Candidate\'s flying hours';
 $string['candidatename'] = 'Candidate name';
+$string['checklistuse'] = 'Use of checklists, airmanship, control of aeroplane by external visual references, anti-icing procedures etc. apply in all sections.';
 $string['comments'] = 'Comments';
 $string['copyright'] = 'Copyright © {$a->ato} {$a->year}. All rights reserved.';
 $string['cumulative'] = 'Cumulative';
@@ -320,13 +326,18 @@ $string['declarationtext'] = 'I certify that the candidate has successfully comp
 $string['dualflight'] = 'Dual fligt instruction';
 $string['duration'] = 'Duration';
 $string['examiner'] = 'Examiner';
-$string['examinerreport'] = 'Examiner Report';
-$string['examinerreportsubject'] = 'This is {$a->ato} ATO Examiner report for the skills test of the {$a->coursename} course.';
+$string['examinerdetails'] = 'Examiner\'s Details';
+$string['examinerreport'] = 'Examiner\'s Skill Test Form';
+$string['examinerreportfor'] = 'Examiner Form for';
+$string['examinerreportsubject'] = 'This is {$a->ato} ATO Examiner form for the skills test of the {$a->coursename} course.';
+$string['examreportcopies'] = 'Copies of the report shall be submitted to (1) The candidate (2) BAVirtual Flight Training (3) The Examiner.';
 $string['recommendationreport'] = 'Recommendation For Skill Test';
 $string['recommendationreportsubject'] = 'This is {$a->ato} ATO Recommendation for Skill Test letter compiled by the instructor recommending a student for a Skill Test.';
 $string['examend'] = 'Exam end';
 $string['examstart'] = 'Exam start';
 $string['flighttraining'] = 'Flight Training';
+$string['flightttest'] = 'Flight Test';
+$string['footnote'] = '*These items may be combined at the discretion of the Examiner.';
 $string['mentorreport'] = 'Mentored Sessions Report';
 $string['mentorreportdesc'] = 'The theory examination is a multi-choice, randomly selected question paper that is time-limited. It is self-contained within the Moodle program for the {$a->ato}
     {$a->coursename} course. {$a->studentname} has completed the test successfully with {$a->attempts} attempt(s) achieving the pass mark of {$a->score} out of {$a->total} ({$a->percent}%) exactly.';
@@ -335,10 +346,14 @@ $string['otherflyingtime'] = 'Other flying time';
 $string['practicalexamreport'] = 'Practical Examination Report';
 $string['practicalexamreportsubject'] = 'This is {$a->ato} ATO Practical Examination report detailing the examiner\'s feedback of the {$a->coursename} course.';
 $string['ratinglabel'] = 'Rating applied for';
-$string['recommendationletterver'] = 'Rev 1 25 Jun 08';
+$string['recommendationletterver'] = 'Rev 2 09 Jun 22';
 $string['recommendedby'] = 'Recommended for skill test by';
+$string['results'] = 'Results';
+$string['skilltestformver'] = 'Rev 3 09 Jun 22';
+$string['testdate'] = 'Date of test';
 $string['theoryexamreport'] = 'Theory Examination Report';
 $string['theoryexamreportsubject'] = 'This is {$a->ato} ATO Theory Examination report summarizing the student\'s score in the {$a->coursename} course.';
+$string['tobecompletedbyexaminer'] = 'to be completed by the examiner';
 $string['trainingaudit'] = 'Flight Training Audit';
 $string['trainingcontent'] = 'Training Content';
 $string['vatsimid'] = 'VATSIM PID';
@@ -346,22 +361,32 @@ $string['vatsimidmissing'] = 'VATSIM PID missing!';
 
 // Skill test form
 $string['aircrafttype'] = "Aircraft Type";
+$string['approvedato'] = "Approved Training Organisation";
 $string['arrivez'] = "Arrive (Z)";
 $string['ato'] = "ATO";
 $string['attempt'] = "Attempt";
 $string['blocktimes'] = "Block times";
 $string['completionconfirmation'] = "I confirm that all required manoeuvres and exercises have been completed";
 $string['departz'] = "Depart (Z)";
+$string['examinerfailsection'] = "Examiner Report: Failure of ";
 $string['fail'] = "Fail";
-$string['furthertraining'] = "Further Training";
+$string['failurereasons'] = "Reasons for Failure";
+$string['furthertraining'] = "Further training";
 $string['itemsnotcompleted'] = "Items not completed";
-$string['pass'] = "Pass";
+$string['mandatory'] = "Mandatory";
+$string['notcompleted'] = "Not Completed";
+$string['notrequired'] = "Not required";
 $string['overallresult'] = "Overall Result";
+$string['pass'] = "Pass";
+$string['recommended'] = "Recommended";
 $string['registration'] = "Registration";
 $string['retestsections'] = "Re-test Sections";
+$string['retestsections'] = "Re-test Sections";
 $string['route'] = "Route";
+$string['section'] = "Section";
 $string['sectionstotake'] = "Sections To Be Taken";
 $string['specifictraining'] = "Specific Training Required";
+$string['subsection'] = "Sub Section";
 $string['testdate'] = "Date of Test";
 $string['testsections'] = "Test Sections";
 $string['testsectionsincomplete'] = "Test sections incomplete due";
@@ -449,11 +474,11 @@ $string['emailinstconfirmhtml'] = '<font face="sans-serif"><p><a href=\'{$a->cou
 $string['emailcancel'] = '{$a->coursename} session booking cancellation: \'{$a->exercise}\'';
 $string['emailcancelmsg'] = '{$a->instructor} has cancelled your booked session scheduled for {$a->sessiondate} on \'{$a->exercise}\'.
 Instructor\'s comment: {$a->comment}.
-Please note you will have to post new availability as current availability posts were purged.';
+Please note that you will have to post new availability slots as previous slots were purged.';
 $string['emailcancelhtml'] = '<font face="sans-serif"><p><a href=\'{$a->courseurl}\'>{$a->coursename}</a> -> <a href=\'{$a->assignurl}\'>Assignment</a> ->
     <a href=\'{$a->exerciseurl}\'>{$a->exercise}</a></p><hr /><p>{$a->instructor} has cancelled your booked session scheduled for
     <strong>{$a->sessiondate}</strong> on \'<i>{$a->exercise}</i>\'.</p><p><strong>Instructor\'s comment:</strong><br />{$a->comment}</p><p>
-    <span style=\'color: red\'>Please note you will have to post new availability as current availability posts were purged.</span></p></p><hr />';
+    <span style=\'color: red\'>Please note that you will have to post new availability slots as previous slots were purged.</span></p></p><hr />';
 
 // email to student: inactive warning
 $string['emailinactivewarning'] = '{$a->coursename}: Inactivity notification';